kernels: shard densities into five translation units
One file held every density instantiation: 227 list entries times every
activity mask times both shapes. It peaked at 7.6 GB of compiler memory
and 150 s of CPU on a fast machine, which is why manylinux builds at
-j1 (two of them at once OOM-killed the runner, exit 137) and why the
mingw build spent 110 of its 122 minutes on this one file while the
other cores idled.

The forwards now live in four shards: densities_common.cpp for the
thirteen tier-3 distributions models lean on plus the hand-written
lpmfs, densities_rest.cpp for the long tail, and two for the 75
distribution functions. densities.cpp keeps registration alone.
densities_impl.hpp holds what they share; the generated forwards moved
from an anonymous namespace into stanli::dens so registration can name
them across translation units, and sum_in_lens had to become inline now
that five TUs see it.

Measured on this machine, -j8, from cold: 180 s to 77 s wall, and peak
RSS 7.57 GB to 4.16. The corpus oracle is the check that matters and it
is unchanged to the character: 119/119, worst hmm_gaussian 3.63e-12,
identical line from both builds. ctest 26/26.

The memory drop is what lets CI raise -j again, which is a separate
change once this has run there.
