+++ runtime/kernels/densities_cdf_a.cpp
@@ -1,0 +1,13 @@
+// Distribution functions, first half.
+//
+// One of the density shards: see densities_impl.hpp for why they
+// are split and what they share.
+#include "densities_impl.hpp"
+
+namespace stanli {
+namespace dens {
+
+STANLI_SCALAR_CDF_LIST_A(STANLI_DEFINE_CDF_FWD)
+
+}  // namespace dens
+}  // namespace stanli
